% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_TempTS.R
\name{bin_TempTS}
\alias{bin_TempTS}
\title{Convert time series to PDT-like data for use in HMMoce.}
\usage{
bin_TempTS(ts, out_dates, bin_res = 8)
}
\arguments{
\item{ts}{Data frame of time series data (Date, Depth, Temperature).}

\item{out_dates}{POSIX vector of desired output dates.}

\item{bin_res}{Numeric indicating desired depth bin resolution (default = 8).}
}
\value{
A dataframe of summarized depth-temperature information.
}
\description{
\code{bin_TempTS} converts time series data (such as from MT tags) to PDT-like data for use in HMMoce
}
\examples{
\dontrun{
tsFile <- system.file("extdata", "141259-Series.csv", package = "HMMoce")
ts <- read.table(tsFile, sep=',', header=T)
ts$Date <- as.POSIXct(paste(ts$Day, ts$Time), format='\%d-\%b-\%Y \%H:\%M:\%S', tz='UTC')
ts <- ts[,c('Date','Depth','Temperature')]

## generate example daily summary of depth-temp profiles
pdt_dates <- seq.POSIXt(tag, pop, by = 'day') 

## generate depth-temp summary from time series
pdt <- bin_TempTS(ts, out_dates = pdt_dates, bin_res = 25)
pdt <- pdt[,c('Date','Depth','MinTemp','MaxTemp')]
}
}
\author{
Ben Galuardi
}
