% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.gpe2.r
\name{calc.gpe2}
\alias{calc.gpe2}
\title{Calculate Light Likelihood from GPE2 Output}
\usage{
calc.gpe2(locs, locDates, locs.grid, dateVec, errEll = TRUE, gpeOnly = TRUE)
}
\arguments{
\item{locs}{is -Locations file output from DAP/Tag Portal for WC tags and 
contains GPS, Argos, and GPE locations as applicable.}

\item{locDates}{is vector of dates from locs dataframe}

\item{locs.grid}{is list output from \code{setup.locs.grid}}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}

\item{errEll}{is logical indicating whether error ellipses should be 
generated for light-based likelihoods as given from output of WC-GPE. False
if only longitude should be used. If False, standard deviation on light 
measurements is currently fixed at 0.7 deg longitude following Musyl et al 
2001. Default is FALSE and will use longitude only.}

\item{gpeOnly}{is logical. If TRUE, locs input is trimmed to GPE positions 
only. This is most applicable in scenarios with FastGPS data and you're 
adding this as a gps input (see "gps" argument).}
}
\value{
L is an array of lon x lat likelihood surfaces (matrices) for each 
  time point (3rd dimension)
}
\description{
\code{calc.gpe2} calculates likelihood estimates for each day of animal tag 
data.
}
\details{
Light errors are parameterized using elliptical error values output in 
'-Locations.csv' (WC tags).
}
\examples{
\dontrun{
# light data as output from GPE2, different filtering algorithm seems to work
# better for light likelihood generation
locs <- read.table(paste(ptt, '-Locations-GPE2.csv', sep = ''), sep = ',', 
                   header = T, blank.lines.skip = F)
L.light <- calc.gpe2(locs, iniloc = iniloc, locs.grid = locs.grid,
                     dateVec = dateVec, errEll = TRUE, gpeOnly = TRUE)
}
}
\seealso{
\code{\link{calc.srss}}
}
