% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.hycom.R
\name{calc.hycom}
\alias{calc.hycom}
\title{Hycom Profile LIkelihood}
\usage{
calc.hycom(
  pdt,
  filename,
  hycom.dir,
  focalDim = 9,
  dateVec,
  use.se = TRUE,
  flip_y = FALSE
)
}
\arguments{
\item{pdt}{input depth-temperature profile data. Need at least cols: Date (POSIXct), Depth, MinTemp, MaxTemp}

\item{filename}{is the first part of the filename specified to the download 
function \code{\link{get.env}}. For example, if downloaded files were 
specific to a particular dataset, you may want to identify that with a name
like 'tuna' or 'shark1'. This results in a downloaded filename of, for 
example, 'tuna_date.nc'. This filename is required here so the calc 
function knows where to get the env data.}

\item{hycom.dir}{directory of downloaded hycom (or other) data}

\item{focalDim}{is integer for dimensions of raster::focal used to calculate 
sd() of temperature grid cell. Recommend focalDim = 9 for Hycom data at 
0.08deg resolution.}

\item{dateVec}{is vector of POSIXct dates for each time step of the likelihood}

\item{use.se}{is logical indicating whether or not to use SE when using 
regression to predict temperature at specific depth levels.}

\item{flip_y}{is logical indicating whether or not to flip the resulting likelihood in the y. Set this to true if output likelihoods are upside down.}
}
\value{
a raster brick of Hycom profile likelihood
}
\description{
Calculate Hycom profile likelihood surface
}
\seealso{
\code{\link{calc.hycom.par}}
}
