% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lightloc.R
\name{calc.lightloc}
\alias{calc.lightloc}
\title{Calculate Light Likelihood from lightloc Output}
\usage{
calc.lightloc(lightloc, locs.grid, dateVec, errEll = TRUE, lon_only = TRUE)
}
\arguments{
\item{lightloc}{is data frame of light-based location estimates. If errEll is FALSE, only Date (POSIXct), Longitude, and Error.Semi.minor.axis (in meters, default output from WC tags) are required. longitudeError (in decimal degrees) can be supplied in place of Error.Semi.minor.axis. If errEll is TRUE, additional required columns are Latitude and Error.Semi.major.axis (meters). These are default outputs from WC tags. Offset (meters) and Offset.orientation (degrees of rotation) are also default outputs from WC tags and are optional columns to include in input. latitudeError (decimal degrees) can be supplied in place of Error.Semi.major.axis. In that case, offset variables are ignored.}

\item{locs.grid}{is list output from \code{setup.locs.grid}}

\item{dateVec}{is vector of POSIXct dates for each time step of the likelihood}

\item{errEll}{is logical indicating whether error ellipses should be 
generated for light-based likelihoods. If FALSE (default), only longitude is used to generate likelihoods. If TRUE, both latitude and longitude are used resulting in an ellipse-shaped likelihood.}

\item{lon_only}{is logical indicating whether the likelihood should be generated for longitude only (default is TRUE). If FALSE, latitude only will be used to generate the likelihood. This argument is only valid is errEll = FALSE.}
}
\value{
L is an array of lon x lat likelihood surfaces (matrices) for each 
  time point (3rd dimension)
}
\description{
\code{calc.lightloc} calculates likelihood estimates for each day of animal tag 
data.
}
\details{
Light errors are parameterized using elliptical error values output in 
'-Locations.csv' (WC tags).
}
\examples{
\dontrun{
## Setup for calculating light likelihood
# Read the data
locsFile <- system.file("extdata", "141259-Locations-GPE2.csv", package = "HMMoce")
locs <- read.table(locsFile, sep = ',', header = TRUE, blank.lines.skip = FALSE)

# Set spatial and temporal limits
sp.lim <- list(lonmin = -82, lonmax = -25, latmin = 15, latmax = 50)
locs.grid <- setup.locs.grid(sp.lim)
iniloc <- data.frame(matrix(c(13, 10, 2015, 41.3, -69.27, 10, 4, 2016, 40.251, -36.061),
 nrow = 2, ncol = 5, byrow = TRUE))
names(iniloc) <- list('day','month','year','lat','lon')
tag <- as.POSIXct(paste(iniloc[1,1], '/', iniloc[1,2], '/', iniloc[1,3], sep=''), 
format = '\%d/\%m/\%Y', tz='UTC')
pop <- as.POSIXct(paste(iniloc[2,1], '/', iniloc[2,2], '/', iniloc[2,3], sep=''), 
format = '\%d/\%m/\%Y', tz='UTC')
dateVec <- as.Date(seq(tag, pop, by = 'day')) 

# Full example light calculation
L.light <- calc.lightloc(locs, iniloc = iniloc, locs.grid = locs.grid,
                     dateVec = dateVec, errEll = TRUE, gpeOnly = TRUE)
}
}
\references{
Musyl MK, Domeier ML, Nasby-Lucas N, Brill RW, McNaughton LM, 
  Swimmer JY, Lutcavage MS, Wilson SG, Galuardi B, Liddle JB (2011) 
  Performance of pop-up satellite archival tags. Mar Ecol Prog Ser
}
\seealso{
\code{\link{calc.srss}}
}
