% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ohc.glorys.par.R
\name{calc.ohc.glorys.par}
\alias{calc.ohc.glorys.par}
\title{Calculate Ocean Heat Content (OHC) likelihood surface}
\usage{
calc.ohc.glorys.par(
  pdt,
  filename,
  isotherm = "",
  ohc.dir,
  dateVec,
  depthVec = NULL,
  bathy = TRUE,
  use.se = TRUE,
  flip_y = FALSE,
  ncores = NULL
)
}
\arguments{
\item{pdt}{input depth-temperature profile data. Need at least cols: Date (POSIXct), Depth, MinTemp, MaxTemp}

\item{filename}{is the first part of the filename specified to the download 
function \code{\link{get.env}}. For example, if downloaded files were 
specific to a particular dataset, you may want to identify that with a name
like 'tuna' or 'shark1'. This results in a downloaded filename of, for 
example, 'tuna_date.nc'. This filename is required here so the calc 
function knows where to get the env data.}

\item{isotherm}{default '' in which isotherm is calculated on the fly based 
on daily tag data. Otherwise, numeric isotherm constraint can be specified 
(e.g. 20 deg C).}

\item{ohc.dir}{directory of downloaded glorys (or other) data}

\item{dateVec}{is vector of POSIXct dates for each time step of the likelihood}

\item{bathy}{is logical indicating whether or not a bathymetric mask should 
be applied. This does NOT use an input bathymetry dataset but rather masks the depth-temperature likelihood based on the possible locations given the maximum recorded tag-depth (i.e. areas with bottom depth shallower than tag-recorded max depth are masked out).}

\item{use.se}{is logical indicating whether or not to use SE when using
regression to predict temperature at specific depth levels.}

\item{flip_y}{is logical indicating whether or not to flip the resulting likelihood in the y. Set this to true if output likelihoods are upside down.}

\item{ncores}{specify number of cores, or leave blank and use whatever you 
have}
}
\value{
likelihood is raster brick of likelihood surfaces representing 
  estimated position based on tag-based OHC compared to calculated OHC using 
  glorys
}
\description{
Compare tag data to OHC grid and calculate likelihoods
}
\references{
Luo J, Ault JS, Shay LK, Hoolihan JP, Prince ED, Brown C a.,
  Rooker JR (2015) Ocean Heat Content Reveals Secrets of Fish Migrations.
  PLoS One 10:e0141101
}
\seealso{
\code{\link{calc.woa}}
}
