% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.pdt.R
\name{extract.pdt}
\alias{extract.pdt}
\title{Extract PDT from Wildlife Computers tag data}
\usage{
extract.pdt(pdt)
}
\arguments{
\item{pdt}{data frame read from -PDTs.csv output of Wildlife Computers DAP 
processor or Tag Portal.}
}
\value{
data frame formatted for pdt data
}
\description{
\code{extract.pdt} is a simple formatting function that parses PDT data and
makes it usable to subsequent functions
}
