% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausskern.isotrop.R
\name{gausskern.isotrop}
\alias{gausskern.isotrop}
\title{Create isotropic Gaussian Kernel}
\usage{
gausskern.isotrop(siz, sigma, muadv = 0, ratio.xy = 1)
}
\arguments{
\item{siz}{size of the kernel, siz x siz. Must be a positive integer.}

\item{sigma}{standard deviation of the kernel. Unit is cell width. Must be a 
positive number.}

\item{muadv}{advection of the kernel. Unit of the input is cell width. 
Defaults to 0.}

\item{ratio.xy}{ratio in x:y}
}
\value{
Gaussian kernel as a 2D matrix of size (siz x siz)
}
\description{
\code{gausskern.isotrop} calculates 2D Gaussian kernel based on kernel size, 
deviation, and advection
}
\note{
No final standardization by the sum. This function is used to design a pseudo-optimal size for the kernel based on sigma while maintaining 99% of the allowable diffusion.
}
\examples{
kern = gausskern.isotrop(3, 0.5)
}
\author{
Paul Gatti
}
