% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mur.sst.R
\name{get.mur.sst}
\alias{get.mur.sst}
\title{Download MUR Sea Surface Temperature (SST) data}
\usage{
get.mur.sst(limits, time, filename = "", download.file = TRUE, dir = getwd())
}
\arguments{
\item{limits}{A list of length 4; minlon, maxlon, minlat, maxlat. Longitude values are -180,180}

\item{time}{A vector of length 2 with the minimum and maximum times in form 
\code{as.Date(date)}.}

\item{filename}{An optional filename. If provided, then the data is 
downloaded to that file. Otherwise the data is not downloaded and the url 
is returned.}

\item{download.file}{Logical. Should use the default \code{download.file} to 
query the server and download or use the optional \code{curl}. Some users 
may need to use \code{curl} in order to get this to work.}

\item{dir}{is local directory where ncdf files should be downloaded to. 
default is current working directory. if enter a directory that doesnt 
exist, it will be created.}
}
\value{
The url used to extract the requested data from the NetCDF subset 
  service.
}
\description{
\code{get.mur.sst} downloads sea surface temperature (SST) data for given
temporal and spatial constraints of your data.
}
\details{
The method may return before the download is completed. It will continue to 
display progress  until the download completes. Change the default 
\code{download.file} if the download is failing on your platform.
}
\examples{
\dontrun{
sp.lim <- list(-90, -60, 0, 30)
time <- as.Date('2013-03-01')
get.oi.sst(sp.lim, time, filename = '')
# only returns url because filename is unspecified
get.oi.sst(sp.lim, time, filename = 'my_data.nc')
nc <- open.nc('my_data.nc')
sst <- var.get.nc(nc, 'sst')
image.plot(sst)
}
  
}
\references{
\url{https://coastwatch.pfeg.noaa.gov/erddap/griddap/jplMURSST41.html}
}
\author{
Function originally written for R by Ben Jones (WHOI) and modified 
  by Camrin Braun and Ben Galuardi.
}
