% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.filter.ext.R
\name{hmm.filter.ext}
\alias{hmm.filter.ext}
\title{HMM filter functions}
\usage{
hmm.filter.ext(g, L, K1, K2, P, maskL = T)
}
\arguments{
\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{final likelihood (2D)}

\item{K1}{first movement (diffusion) kernel see \code{\link{gausskern}}}

\item{K2}{second movement (diffusion) kernel see \code{\link{gausskern}}}

\item{P}{2x2 probability matrix for tranisitons between states (K1 and K2)}

\item{maskL}{is logical indicating whether to mask the input L layer. See
\code{mask.L} for details.}
}
\value{
a list: list(phi = phi, pred = pred, psi = psi) where \itemize{ \item
  phi. is the probability for each state at each trime step \item pred. is
  .... \item psi. is.... }
}
\description{
HMM filter functions
}
\examples{
\dontrun{

# RUN THE FILTER STEP
f <- hmm.filter(g, L, K1, K2, P.final)

}

}

