% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.filterExp.R
\name{hmm.filterExp}
\alias{hmm.filterExp}
\title{HMM filter functions (modified for mini-expansion method)}
\usage{
hmm.filterExp(
  g,
  L,
  K,
  sigmas,
  n.div = 10,
  bathy,
  P,
  m,
  maskL = FALSE,
  bound.thr = 0.1,
  minBounds = 10,
  mini.idx = NULL
)
}
\arguments{
\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{is likelihood array output from \code{make.L}}

\item{K}{is list of length m where each element represents the movement (diffusion) kernel for that behavior state \code{\link{gausskern}}}

\item{sigmas}{is vector of length m where each element represents the one-dimensional variance for diffusion of the behavior state(s)}

\item{n.div}{is the number of mini-expansion steps the original diffusion kernel is divided into (default = 10, but test carefully for each use case)}

\item{bathy}{is the bathymetry raster used in landmasking - must be the same resolution and size as your grid (recommend bathy_resamp)}

\item{P}{a matrix of dimensions m x m representing the probability for transitions between states.}

\item{m}{is integer indicating the number of desired behavior states. Currently limited to max 2.}

\item{maskL}{is logical indicating whether to mask the input L layer. Default is FALSE. See
\code{mask.L} for details.}

\item{bound.thr}{is numeric indicating the percent threshold that is added 
and subtracted from the bounding box of the filter output from the 
previous day before masking. Default is .05 (5 percent).}

\item{minBounds}{is size (in degrees) of the minimum bounding box around the 
previous days filter prediction that L data within that box will be 
included. Outside this box (centered on t-1 filter prediction), L will be 
masked out.}

\item{mini.idx}{is a numeric vector that corresponds to the positions of dateVec for which the mini-expansion should be applied.}
}
\value{
a list: list(phi = phi, pred = pred, psi = psi) where \itemize{ \item
  phi. is the probability for each state at each time step \item pred. is
  .... \item psi. is.... }
}
\description{
Original expanding kernel code by Dr. Julie Nielsen, integrated by Dr. Martin Arostegui (aka Dr. Mola) into pre-existing hmm.filter function in HMMoce
}
\details{
Cite: Nielsen JK, Bryan DR, Rand KM, Arostegui MC, Braun CD, Galuardi B, McDermott SF. YYYY. Geolocation of a demersal fish (Pacific cod) in a high-latitude island chain (Aleutian Islands, Alaska). Animal Biotelemetry VV: PP-PP.
}
\examples{
\dontrun{
# Not run as function relies on large arrays of likelihoods
# RUN THE FILTER STEP
f <- hmm.filter(g, L, K = list(K1, K2), P = P.final, m = 2)
nllf <- -sum(log(f$psi[f$psi>0])) # negative log-likelihood

}

}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}
