% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.posteriorviterbi.R
\name{hmm.posteriorviterbi}
\alias{hmm.posteriorviterbi}
\title{Collaborative development by Drs. Julie Nielsen, Martin Arostegui, Camrin Braun, and Ben Galuardi
Cite: 
Nielsen JK, Bryan DR, Rand KM, Arostegui MC, Braun CD, Galuardi B, McDermott SF (2023) Geolocation of a demersal fish (Pacific cod) in a high-latitude island chain (Aleutian Islands, Alaska). Animal Biotelemetry 11: 29.}
\usage{
hmm.posteriorviterbi(
  pars,
  s,
  g,
  iniloc,
  dateVec,
  lims.lat,
  lims.lon,
  percentile = 99.9,
  approach = "separate"
)
}
\arguments{
\item{pars}{HMM parameters output from opt.params()}

\item{s}{HMM backward smoothed output of hmm.smoother()}

\item{g}{original grid on which HMM was run, assigned from L.res$g}

\item{iniloc}{dataframe of tag and popup locations/dates (2 rows, 5 columns)}

\item{dateVec}{timestep sequence used in HMM}

\item{lims.lat}{vector of c(min,max) desired latitude for reduced spatial extent in which to run algorithm}

\item{lims.lon}{vector of c(min,max) desired longitude for reduced spatial extent in which to run algorithm}

\item{percentile}{uncertainty percentile of the daily posterior to include in admissible paths (default is 99.9% of daily posterior)}

\item{approach}{for two-state outputs, use a priori most likely behavior state sequence with separate posteriors and kernels ('separate'), or joint posteriors and kernels weighted by relative behavior state probability ('joint')}
}
\description{
Parameters
}
