% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.smoother.r
\name{hmm.smoother}
\alias{hmm.smoother}
\title{Smoother recursion over filtered state estimates}
\usage{
hmm.smoother(f, K1, K2, P)
}
\arguments{
\item{f}{is array output from \code{hmm.filter}}

\item{K1}{is movement kernel generated by \code{gausskern} for behavior state
1}

\item{K2}{is movement kernel generated by \code{gausskern} for behavior state
2}

\item{P}{is transition matrix (usually 2x2) representing probability of state
switching}
}
\value{
an array of the final state estimates of dim(state, time, lon, lat)
}
\description{
\code{hmm.smoother} provides backward (starting at end) recursion over 
filtered state estimates as output from \code{hmm.filter}. The product of 
this function an array containing final state estimates.
}

