% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.smootherExp.R
\name{hmm.smootherExp}
\alias{hmm.smootherExp}
\title{Smoother recursion over filtered state estimates}
\usage{
hmm.smootherExp(f, L, K, P, bathy, n.div, mini.idx, sigmas)
}
\arguments{
\item{f}{is array output from \code{hmm.filter}}

\item{L}{is likelihood array output from \code{make.L}}

\item{K}{is list of length m where each element represents the movement (diffusion) kernel for that behavior state \code{\link{gausskern}}}

\item{P}{a matrix of dimensions m x m representing the probability for transitions between states.}
}
\value{
an array of the final state estimates of dim(state, time, lon, lat)
}
\description{
\code{hmm.smoother} provides backward (starting at end) recursion over 
filtered state estimates as output from \code{hmm.filter}. The product of 
this function an array containing final state estimates.
}
\examples{
\dontrun{
# Not run as function relies on large arrays of likelihoods
# RUN THE SMOOTHING STEP
s <- hmm.smoother(f, K1, K2, L, P.final)
}

}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}
