% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregular_ncToRaster.R
\name{irregular_ncToRaster}
\alias{irregular_ncToRaster}
\title{Coerce irregular grid as nc to raster}
\usage{
irregular_ncToRaster(fname, varid)
}
\arguments{
\item{fname}{is input filename of the .nc file}

\item{varid}{is character indicating the name of the variable of interest in the input nc variable}
}
\value{
a raster version of the input .nc file coerced to regular grid
}
\description{
Main purpose of this function is for converting bathymetry data that is typically acquired as an irregular grid
}
\examples{
\dontrun{
bathy <- irregular_ncToRaster(paste0(bathy.dir, 'bathy.nc'), varid = 'topo')
}
}
