% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liksrss.R
\name{liksrss}
\alias{liksrss}
\title{Calculate density function and integrate between limits}
\usage{
liksrss(obs, srss, srsd)
}
\arguments{
\item{obs}{input light observation. See \code{\link{calc.srss}}}

\item{srss}{is raster of possible SR or SS times within study area}

\item{srsd}{is raster of SD of SRSS times in study area calculated with \code{raster::focal}}
}
\value{
an array of SRSS-based likelihoods
}
\description{
#' \code{liksrss} calculates density function for a normal distribution and
integrates between limits of SRSS times (dawn/dusk)
}
