% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.L.pdt.R
\name{make.L.pdt}
\alias{make.L.pdt}
\title{Combine individual source PDT likelihoods}
\usage{
make.L.pdt(ras.list, dateVec)
}
\arguments{
\item{ras.list}{a list of likelihood rasters}

\item{dateVec}{is vector of POSIXct dates for each time step of the likelihood}
}
\value{
an overall PDT-based likelihood RasterStack
}
\description{
\code{make.L.pdt} combines individual PDT-based likelihoods from various data sources (e.g.
HYCOM OHC, HYCOM depth-levels, GLORYS OHC, GLORYS depth-levels) to make overall 
combined PDT-based likelihoods for each time point
}
\examples{
\dontrun{
L <- make.L(ras.list,dateVec)
}
}
\author{
Martin C. Arostegui
}
