% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.L.R
\name{make.L}
\alias{make.L}
\title{Combine individual source likelihoods}
\usage{
make.L(ras.list, iniloc, dateVec, maxDepth, bathy, known.locs = NULL)
}
\arguments{
\item{ras.list}{a list of likelihood rasters}

\item{iniloc}{is data.frame of tag and pop locations. Required columns are 'date' (POSIX), 'lon', 'lat'.}

\item{dateVec}{is vector of POSIXct dates for each time step of the likelihood}

\item{maxDepth}{a vector of the daily max depths (must be positive and be zero-padded for NA days)}

\item{bathy}{is the original bathy raster pre-resampling}

\item{known.locs}{is data frame of known locations containing named columns 
of 'date' (POSIX), 'lon', 'lat'. Default is NULL.}
}
\value{
an overall likelihood array, L
}
\description{
\code{make.L} combines individual likelihoods from various data sources (e.g.
SST, OHC) to make overall combined likelihoods for each time point
}
\examples{
\dontrun{
L <- make.L(ras.list, iniloc, dateVec)
}
}
