% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make180.R
\name{make180}
\alias{make180}
\title{convert  0, 360 --> -180, 180}
\usage{
make180(lon)
}
\arguments{
\item{lon}{is numeric or integer longitude to be converted}
}
\value{
converted lon
}
\description{
convert  0, 360 --> -180, 180
}
