% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make360.R
\name{make360}
\alias{make360}
\title{convert -180, 180 --> 0, 360}
\usage{
make360(lon)
}
\arguments{
\item{lon}{is numeric or integer longitude to be converted}
}
\value{
converted lon
}
\description{
convert -180, 180 --> 0, 360
}
