% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microwave_to_etuff.R
\name{microwave_to_etuff}
\alias{microwave_to_etuff}
\title{Convert Microwave Telemetry XLS/XLSX to eTUFF file}
\usage{
microwave_to_etuff(excel_file, meta_row, out_file = NULL, obs_types_url = NULL)
}
\arguments{
\item{excel_file}{Path to Microwave .xls or .xlsx file}

\item{meta_row}{Data frame containing tag metadata (instrument_name, platform, etc.)}

\item{out_file}{Destination path for the .txt eTUFF file}

\item{obs_types_url}{Optional custom URL/path to eTUFF-ObservationTypes.csv}
}
\value{
An object of class \code{etuff}
}
\description{
Convert Microwave Telemetry XLS/XLSX to eTUFF file
}
