% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg.log.lik.fun.R
\name{neg.log.lik.fun}
\alias{neg.log.lik.fun}
\title{Calculate negative log likelihood}
\usage{
neg.log.lik.fun(
  pars,
  g,
  L,
  maskL = TRUE,
  bound.thr = 0.1,
  minBounds = 10,
  kernel = "normal"
)
}
\arguments{
\item{pars}{kernel parameters. Usually output from \code{est_params}}

\item{g}{see \code{hmm.filter}}

\item{L}{see \code{hmm.filter}}

\item{maskL}{see \code{hmm.filter}}

\item{bound.thr}{see \code{hmm.filter}}

\item{minBounds}{see \code{hmm.filter}}

\item{kernel}{character either "normal" or "isotropic", depending on desired kernel characteristics}
}
\value{
a negative log likelihood
}
\description{
\code{neg.log.lik.fun} calculates the negative log likelihood
}
\author{
Paul Gatti
}
