% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.params.R
\name{opt.params}
\alias{opt.params}
\title{Estimate parameters}
\usage{
opt.params(
  pars.init,
  lower.bounds,
  upper.bounds,
  g,
  L,
  alg.opt = "optim",
  reformat = TRUE,
  ...
)
}
\arguments{
\item{pars.init}{is vector of initial parameter set: sigma1 (resident), sigma2 (migrant), p11 switch from behavior 1 to 1 (remain resident), p22 switch from behavior 2 to 2 (remain migrant)}

\item{lower.bounds}{vector of lower bounds of pars.init}

\item{upper.bounds}{vector of upper bounds of pars.init}

\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{is likelihood array output from \code{make.L}}

\item{alg.opt}{character indicating which optimization algorithm to use. Options are "optim", "nlminb", "ga", "ga_optim".}

\item{reformat}{logical indicating whether to reformat and thus standardize output of the optimization algorithms}

\item{write.results}{optional. logical indicating whether to write results with \code{saveRDS}.}

\item{hessian}{optional. logical indicating whether to compute hessian matrix for estimate of parameter standard deviation. Note: usually slows down the computations significantly. Default is FALSE.}

\item{maskL}{optional. logical indicating whether to apply likelihood mask in \code{hmm.filter}. Default is FALSE.}

\item{optim_method}{optional. see \code{method} input to \code{optim}. Default is "L-BFGS-B".}

\item{filename}{optional. if \code{write.results == TRUE}, this should satisfy \code{file} argument to \code{saveRDS}}

\item{max_iter}{optional. see \code{maxiter} input to \code{GA::ga()}. Default is 100.}

\item{run}{optional. see \code{run} input to \code{GA::ga()}. Default is 100.}

\item{p_size}{optional. see \code{popSize} input to \code{GA::ga()}. Default is 100.}

\item{ncores}{optional. see \code{parallel} input to \code{GA::ga()}. Default is FALSE, therefore not parallelized.}
}
\description{
\code{opt.params} is a wrapper for several different parameter optimization routines
}
\examples{
\dontrun{
pars.init=c(2,.2,.6,.8)
lower.bounds=c(0.1,0.001,.1,.1)
upper.bounds=c(5,.5,.9,.9)

}
}
\author{
Paul Gatti

Camrin Braun
}
