% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.log.lik.fun.R
\name{pos.log.lik.fun}
\alias{pos.log.lik.fun}
\title{Calculate positive log likelihood}
\usage{
pos.log.lik.fun(
  pars,
  g,
  L,
  maskL = TRUE,
  bound.thr = 0.1,
  minBounds = 10,
  kernel = "normal"
)
}
\arguments{
\item{pars}{kernel parameters. Usually output from \code{est_params}}

\item{g}{see \code{hmm.filter}}

\item{L}{see \code{hmm.filter}}

\item{maskL}{see \code{hmm.filter}}

\item{bound.thr}{see \code{hmm.filter}}

\item{minBounds}{see \code{hmm.filter}}

\item{kernel}{character either "normal" or "isotropic", depending on desired kernel characteristics}
}
\value{
a positive log likelihood
}
\description{
\code{pos.log.lik.fun} calculates the positive log likelihood
}
\details{
For use in \code{GA::ga()} algorithm which maximizes, instead of minimizes, the cost/objective function
}
\author{
Paul Gatti
}
