% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.grid.par.r
\name{resample.grid.par}
\alias{resample.grid.par}
\title{Resample likelihood rasters to common resolution/extent}
\usage{
resample.grid.par(L.rasters, L.resol, ncores = NULL)
}
\arguments{
\item{L.rasters}{list of individual likelihood rasters generated by calc 
functions}

\item{L.resol}{raster or raster brick indicating desired output resolution of
all likelihood rasters.}

\item{ncores}{is integer indicating number of cores used in this parallel
computation. Defaults to using a detection function that chooses cores for
you.}
}
\value{
a list of all resampled likelihood rasters and g, the common grid
}
\description{
Resample likelihood rasters to common resolution/extent
}
\examples{
\dontrun{
 # use with caution as all workspace items containing 'L.' will be listed. 
 # We only want the likelihood outputs calculated above
L.rasters <- mget(ls(pattern = 'L\\\\.'))
resamp.idx <- which.max(lapply(L.rasters, FUN=function(x) raster::res(x)[1]))
L.res <- resample.grid.par(L.rasters, L.rasters[[resamp.idx]])
}
}
