% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.grid.R
\name{resample.grid}
\alias{resample.grid}
\title{Resample likelihood rasters to common resolution/extent}
\usage{
resample.grid(L.rasters, L.res, aggregate = FALSE, cellAgg = NULL)
}
\arguments{
\item{L.rasters}{list of individual likelihood rasters generated by calc 
functions}

\item{L.res}{raster or raster brick indicating desired output resolution of
all likelihood rasters.}

\item{aggregate}{is logical indicating whether you want the function to try
to aggregate grid cells to generate a coarse raster for parameter
estimation that ends up around resolution of 1deg. Otherwise, the coarse
grid outputs will be based on the max resolution of inputs rasters.}

\item{cellAgg}{is fact argument to \code{\link[raster]{aggregate}}}
}
\value{
a list of all resampled likelihood rasters and g, the common grid
}
\description{
Resample likelihood rasters to common resolution/extent
}
