% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.grid.raster.R
\name{setup.grid.raster}
\alias{setup.grid.raster}
\title{Setup the discrete spatial grid for the HMM}
\usage{
setup.grid.raster(grid.ras)
}
\arguments{
\item{grid.ras}{is a raster for which a grid is desired}
}
\value{
a list
}
\description{
Setup the discrete spatial grid for the HMM
}
