% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.grid.R
\name{setup.grid}
\alias{setup.grid}
\title{Setup the discrete spatial grid for the HMM}
\usage{
setup.grid(locations, res)
}
\arguments{
\item{locations}{dataframe of -Locations from WC psat tag}

\item{res}{character indicating resolution of grid. 'hycom' is .08 to match
hycom reanalysis res. 'quarter' and 'one' are .25 and 1 deg, respectively.}
}
\value{
a list
}
\description{
Setup the discrete spatial grid for the HMM
}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}
