% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_TempTS.R
\name{bin_TempTS}
\alias{bin_TempTS}
\title{Convert time series to PDT-like data for use in HMMoce.}
\usage{
bin_TempTS(ts, tsDates = NULL, res = 8)
}
\arguments{
\item{ts}{is data frame of time series data. Currently this is built for WC 
tags but can handle others if you know what youre doing (e.g. proper input 
format). Required inputs cols are just date, Depth and Temperature.}

\item{tsDates}{is vector of desired dates as output. This is designed to be
used for creating higher temporal resolution PDT datasets from time series
data.}

\item{res}{is numeric indicating desired depth bin resolution. Default is 8 
to match native PDT format on WC tags; however, theres no reason why this 
couldnt be higher resolution, especially if consider we later match this to
a 3D product such as HYCOM that is higher depth resolution (usually). **NOT WORKING**}
}
\value{
a dataframe that matches the required format for a PDT data product 
  for further use in HMMoce
}
\description{
\code{bin_TempTS} compares tag SST to remotely sensed SST and calculates 
likelihoods
}
\examples{
\dontrun{
}
}
