% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.bottomTemp.r
\name{calc.bottomTemp}
\alias{calc.bottomTemp}
\title{Calculate bt-based likelihood}
\usage{
calc.bottomTemp(tag.bt, filename, bt.dir, dateVec, focalDim = NULL,
  sens.err = 1, varName = "Temperature")
}
\arguments{
\item{tag.bt}{is data frame containing tag-collected bottom temperature data. Requires at least cols: Date (POSIXct), Temperature}

\item{filename}{is the first part of the filename specified to the download 
function \code{\link{get.env}}. For example, if downloaded files were 
specific to a particular dataset, you may want to identify that with a name
like 'tuna' or 'shark1'. This results in a downloaded filename of, for 
example, 'tuna_date.nc'. This filename is required here so the calc 
function knows where to get the env data.}

\item{bt.dir}{local directory where bottom temp nc grids are stored}

\item{dateVec}{is vector of POSIXct dates for each time step of the likelihood}

\item{focalDim}{is integer for dimensions of raster::focal used to calculate 
sd() of env grid cell. If left to NULL (default), this dimension will
approximately incorporate 0.25 degrees.}

\item{sens.err}{is numeric indicating the percent sensor error in the tag bottom temp
sensor. Default is 1.}

\item{varName}{is name of temperature variable in the bottom temperature nc grids. Default is "Temperature".}
}
\value{
likelihood is raster brick of likelihood surfaces representing
  matches between tag-based bottom temperature and environmental bottom temperature grids
}
\description{
\code{calc.bottomTemp} compares tag-measured bottom temperature to a bottom temperature grid and calculates 
likelihoods
}
\seealso{
\code{\link{calc.sst}}
}
