% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.errEll.r
\name{calc.errEll}
\alias{calc.errEll}
\title{Calculates error ellipses around input locations}
\usage{
calc.errEll(locs, locs.grid)
}
\arguments{
\item{locs}{is data frame of locations typically read from -Locations.csv output
from Wildlife Computers}

\item{locs.grid}{is list output from \code{setup.locs.grid}}
}
\value{
an array of error ellipses for input light-based locations
}
\description{
Error ellipses are generated around WC tag based positions using
tag-calculated error structure.
}
