% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.hycom.par.r
\name{calc.hycom.par}
\alias{calc.hycom.par}
\title{Hycom Profile LIkelihood in Parallel}
\usage{
calc.hycom.par(pdt, ptt, hycom.dir, focalDim = 9, dateVec,
  ncores = detectCores())
}
\arguments{
\item{pdt}{input PAT data see \code{\link{extract.pdt}}}

\item{ptt}{name of tag i.e. 123645}

\item{hycom.dir}{directory of downloaded hycom (or other)data}

\item{focalDim}{is integer for dimensions of raster::focal used to calculate
sd() of temperature grid cell. Recommend focalDim = 3 if woa.data = woa.one
and 9 if using woa.quarter.}

\item{dateVec}{vector of complete dates for data range. This should be in 'Date' format}

\item{ncores}{specify number of cores, or leave blank and use whatever you have!}
}
\value{
a raster brick of Hycom profile likelihood
}
\description{
Calculate Hycom profile probability surface in parallel
}
\examples{
# load workspace
\dontrun{
load('~/DATA/blue259.RData')
# define hycom.dir
hycom.dir = '~/hycom/'
# run in parallel
res = calc.hycom.par(pdt, ptt, isotherm = '', hycom.dir = hycom.dir, dateVec = dateVec, bathy = T)
}
}

