% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.hycom.r
\name{calc.hycom}
\alias{calc.hycom}
\title{Hycom Profile LIkelihood in Parallel}
\usage{
calc.hycom(pdt, filename, hycom.dir, focalDim = 9, dateVec, use.se = TRUE)
}
\arguments{
\item{pdt}{input PAT data see \code{\link{extract.pdt}}}

\item{filename}{is the first part of the filename specified to the download 
function \code{\link{get.env}}. For example, if downloaded files were 
specific to a particular dataset, you may want to identify that with a name
like 'tuna' or 'shark1'. This results in a downloaded filename of, for 
example, 'tuna_date.nc'. This filename is required here so the calc 
function knows where to get the env data.}

\item{hycom.dir}{directory of downloaded hycom (or other)data}

\item{focalDim}{is integer for dimensions of raster::focal used to calculate 
sd() of env grid cell. If left to NULL (default), this dimension will 
approximately incorporate 0.25 degrees.}

\item{dateVec}{vector of complete dates for data range. This should be in 
'Date' format}

\item{use.se}{is logical indicating whether or not to use SE when using 
regression to predict temperature at specific depth levels.}
}
\value{
a raster brick of Hycom profile likelihood
}
\description{
Calculate Hycom profile probability surface in parallel
}
\examples{
# load workspace
\dontrun{
load('~/DATA/blue259.RData')
# define hycom.dir
hycom.dir = '~/hycom/'
# run in parallel
res = calc.hycom.par(pdt, filename='tuna', isotherm = '',
                     hycom.dir = hycom.dir, dateVec = dateVec,
                     bathy = T)
}
}
