% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ohc.r
\name{calc.ohc}
\alias{calc.ohc}
\title{Calculate Ocean Heat Content (OHC) probability surface}
\usage{
calc.ohc(pdt, ptt, isotherm = "", ohc.dir, dateVec, bathy = TRUE,
  use.se = TRUE)
}
\arguments{
\item{pdt}{is variable containing tag-collected PDT data}

\item{ptt}{is unique tag identifier}

\item{isotherm}{default '' in which isotherm is calculated on the fly based 
on daily tag data. Otherwise, numeric isotherm constraint can be specified 
(e.g. 20 deg C).}

\item{ohc.dir}{local directory where \code{get.env} downloads are stored.}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}

\item{bathy}{is logical indicating whether or not a bathymetric mask should
be applied}

\item{use.se}{is logical indicating whether or not to use SE when using regression to predict temperature at specific depth levels.}
}
\value{
likelihood is raster brick of likelihood surfaces representing 
  estimated position based on tag-based OHC compared to calculated OHC using 
  HYCOM
}
\description{
Compare tag data to OHC grid and calculate likelihoods
}
\examples{
\dontrun{
# depth-temp profile data
pdt <- read.wc(ptt, wd = myDir, type = 'pdt', tag=tag, pop=pop); 
pdt.udates <- pdt$udates; pdt <- pdt$data
# GENERATE DAILY OCEAN HEAT CONTENT (OHC) LIKELIHOODS
L.ohc <- calc.ohc(pdt, ptt, ohc.dir = hycom.dir, dateVec = dateVec,
                  isotherm = '')
}
}
\seealso{
\code{\link{calc.woa}}
}

