% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.sst.par.r
\name{calc.sst.par}
\alias{calc.sst.par}
\title{Calculate SST-based likelihood}
\usage{
calc.sst.par(tag.sst, ptt, sst.dir, dateVec, sens.err = 1,
  ncores = parallel::detectCores())
}
\arguments{
\item{tag.sst}{variable containing tag-collected SST data}

\item{ptt}{is unique tag identifier}

\item{sst.dir}{local directory where remote sensing SST downloads are stored}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}

\item{sens.err}{is numeric indicating the percent sensor error in the tag sst sensor. Default is 1.}

\item{ncores}{is integer indicating number of cores used in this parallel computation. Defaults to using a detection function that chooses cores for you.}
}
\value{
likelihood is raster brick of likelihood surfaces representing matches
  between tag-based sst and remotely sensed sst maps
}
\description{
\code{calc.sst.par} compares tag SST to remotely sensed SST and calculates
likelihoods
}
\examples{
\dontrun{

# sst data
tag.sst <- read.wc(ptt, wd = myDir, type = 'sst', tag=tag, pop=pop); 
sst.udates <- tag.sst$udates; tag.sst <- tag.sst$data

# IF YOU NEED TO DOWNLOAD SST DATA
sst.dir <- paste('my_sst_dir')
get.env(sst.udates, type = 'sst', spatLim = sp.lim, save.dir = sst.dir)

# GENERATE DAILY SST LIKELIHOODS
L.sst <- calc.sst(tag.sst, sst.dir = sst.dir, dateVec = dateVec)

}
}
\seealso{
\code{\link{calc.ohc}}
}

