% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.track.r
\name{calc.track}
\alias{calc.track}
\title{Calculate movement track from state estimates}
\usage{
calc.track(distr, g)
}
\arguments{
\item{distr}{is output array from \code{hmm.smoother}}

\item{g}{is one of the outputs from \code{resample.grid} which denotes what
spatial scale and grid you're working on}
}
\value{
calculated track
}
\description{
Calculate movement track from state estimates
}

