% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.track.r
\name{calc.track}
\alias{calc.track}
\title{Calculate movement track from state estimates}
\usage{
calc.track(distr, g, dateVec, method = "mean")
}
\arguments{
\item{distr}{is output array from \code{hmm.smoother}}

\item{g}{is one of the outputs from \code{resample.grid} which denotes what 
spatial scale and grid you're working on}

\item{dateVec}{is vector of dates from tag to pop-up in 1 day increments.}

\item{method}{is character indicating what method to use for track 
calculation. Currently only 'mean' is supported.}
}
\value{
calculated track
}
\description{
\code{calc.track} uses HMM output via \code{hmm.smoother} to calculate most
probable track and behavior state
}
\examples{
\dontrun{

# RUN THE SMOOTHING STEP
s = hmm.smoother(f, K1, K2, P.final)

# GET AND PLOT THE MOST PROBABLE TRACK
tr <- calc.track(s, g, dateVec)

}

}

