% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse.L.r
\name{coarse.L}
\alias{coarse.L}
\title{Create coarse overall likelihood for efficient parameter estimation}
\usage{
coarse.L(L, ras.list, aggregate_fact = NULL)
}
\arguments{
\item{L}{is overall likelihood array output from \code{make.L}}

\item{ras.list}{a list of likelihood rasters as input to \code{make.L}. Can also be just one of these raster layers rather than of class "list".}

\item{aggregate_fact}{is \code{fact} input to \code{raster::aggregate}.}
}
\description{
\code{coarse.L} makes a more coarse (aggregated) version of the overall likelihood array (output from \code{make.L}). This results in a coarse version of the overall likelihood grid used for more efficient parameter estimation.
}
