% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expmax.r
\name{expmax}
\alias{expmax}
\title{Expectation-maximization framework for state-switching}
\usage{
expmax(p.init, g, L, K1, K2, niter = 1000, threshold = 0.01)
}
\arguments{
\item{p.init}{a vector of length 2. The first is the probability of staying 
in behavior state 1 if currently in state 1 and the second for staying in 
state 2.}

\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{final likelihood}

\item{K1}{first movement (diffusion) kernel see \code{\link{gausskern}}}

\item{K2}{second movement (diffusion) kernel see \code{\link{gausskern}}}

\item{niter}{is integer that determines number of iterations to perform}

\item{threshold}{is threshold of percent change that we consider satisfactory
for convergence. Default is 1\%.}
}
\value{
a 2x2 matrix of state switching probabilities. See P.init input for 
  more information.
}
\description{
\code{expmax} performs expectation-maximization for state switching 
probability
}
\details{
Light errors are parameterized using elliptical error values output in.
}
\references{
Woillez M, Fablet R, Ngo TT, et al. (2016) A HMM-based model to
  geolocate pelagic fish from high-resolution individual temperature and
  depth histories: European sea bass as a case study. Ecol Modell 321:10-22.
}

