% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausskern.nostd.r
\name{gausskern.nostd}
\alias{gausskern.nostd}
\title{Create Gaussian Kernel}
\usage{
gausskern.nostd(siz, sigma, muadv = 0)
}
\arguments{
\item{siz}{size of the kernel, siz x siz. Must be a positive integer.}

\item{sigma}{standard deviation of the kernel. Unit is cell width. Must be a 
positive number.}

\item{muadv}{advection of the kernel. Unit of the input is cell width. 
Defaults to 0.}
}
\value{
Gaussian kernel as a 2D matrix of size (siz x siz)
}
\description{
\code{gausskern.nostd} calculates 2D Gaussian kernel based on kernel size, 
deviation, and advection. Differs from \code{gausskern} in that it should result in a more circular, and perhaps realistic, kernel. This *should* avoid the issue with square kernels allowing more diffusion in the diagonal than horizontal or vertical.
}
\examples{
kern = gausskern.nostd(3, 0.5)

}
\author{
Paul Gatti
}
