% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bath.data.r
\name{get.bath.data}
\alias{get.bath.data}
\title{Download bathymetry data}
\usage{
get.bath.data(spatLim, save.dir = tempdir(), seaonly = TRUE, res = c(0.5))
}
\arguments{
\item{spatLim}{is a list of spatial limits as \code{list(lonmin, lonmax, latmin, latmax)}}

\item{save.dir}{is destination folder. Default is a temporary directory.}

\item{seaonly}{is logical indicating whether you want to download only
bathymetry below sea level.}

\item{res}{is numeric indicating resolution in minutes. Choices currently are 0.5 or 1minute.}
}
\value{
Downloads a NetCDF file containing ETopo bathymetry and also returns the output as a \code{raster} object.
}
\description{
Download ETOPO bathymetry. Resolution is either 30 second or one minute 
resolution
}
\note{
Be patient! The download can take a few minutes!
}
\examples{
\dontrun{
# Not run to prevent actual data download
sp.lim <- list(lonmin = -82, lonmax = -25, latmin = 15, latmax = 50)
bathy <- get.bath.data(sp.lim, save.dir = tempdir())
}
}
\author{
Function originally written by Ben Galuardi.
}
