% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bath.data.r
\name{get.bath.data}
\alias{get.bath.data}
\title{Download bathymetry
res is either 30 second, one or two minute resolution (.5, 1)}
\usage{
get.bath.data(lonlow, lonhigh, latlow, lathigh, folder = tempdir(),
  seaonly = T, res = c(0.5), raster = TRUE)
}
\arguments{
\item{lonlow}{numeric indicating minimum longitude extent of desired download (-180 to 180).}

\item{lonhigh}{see lonlow}

\item{latlow}{see lonlow}

\item{lathigh}{see lonlow}

\item{folder}{is destination folder. Default is a temporary directory.}

\item{seaonly}{is logical indicating whether you want to mask out anything on land.}

\item{res}{is numeric indicating resolution in minutes. Choices currently limited to just 0.5.}

\item{raster}{is logical indicating whether you want the function to return a raster or not (a list will be returned).}
}
\description{
Download bathymetry
res is either 30 second, one or two minute resolution (.5, 1)
}
\note{
Be patient! The download can take a few minutes!
}
