% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.env.r
\name{get.env}
\alias{get.env}
\title{Download and Read Oceanographic Data}
\usage{
get.env(uniqueDates = NULL, ptt = NULL, type = NULL, spatLim = NULL,
  resol = NULL, save.dir = getwd(), sst.type = NULL)
}
\arguments{
\item{uniqueDates}{is a POSIXct vector of desired dates}

\item{ptt}{is unique tag identifier}

\item{type}{is a character string indicating whether you're after sea surface
temperature 'sst', hybrid coordinate ocean model 'hycom', or world ocean atlas 'woa' data}

\item{spatLim}{is a list of spatial limits as \code{list(xmin, xmax, ymin,
ymax)}}

\item{resol}{is character describing the desired resolution in degrees if type = 'woa', otherwise NULL.}

\item{save.dir}{is the directory to save the downloaded data to}
}
\value{
nothing, just downloads the data to your local machine
}
\description{
\code{get.env} accesses oceanographic data like sea surface temperature from
a remote server and downloads the temporal and spatial extent of interest for
further use
}
