% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nll.fun.r
\name{get.nll.fun}
\alias{get.nll.fun}
\title{Negative Log likelihood of parameters}
\usage{
get.nll.fun(parvec = c(10, 30, 5, 2, 0.707, 0.8), g, L)
}
\arguments{
\item{parvec}{vector of length 6 containing * kernel 1 * kernel 2 * diagonal
of 2x2 matrix}

\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{final likelihood (2D)}
}
\value{
parameter values
}
\description{
Negative Log likelihood of parameters
}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}
