% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.woa.r
\name{get.woa}
\alias{get.woa}
\title{Download World Ocean Atlas Climatology}
\usage{
get.woa(save.dir = getwd(), resol = "one")
}
\arguments{
\item{save.dir}{is the directory to save the downloaded data to}

\item{resol}{is character describing the desired resolution in degrees. Choices are 'one' or 'quarter'.}
}
\value{
name and directory of the downloaded data
}
\description{
\code{get.woa} downloads World Ocean Atlas 2013 mean climatological
temperature data from GitHub
}
