% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.filter.r
\name{hmm.filter}
\alias{hmm.filter}
\title{HMM filter functions}
\usage{
hmm.filter(g, L, K, P, m, maskL = FALSE, bound.thr = 0.1,
  minBounds = 10)
}
\arguments{
\item{g}{grid from \code{\link{setup.grid}}}

\item{L}{is likelihood array output from \code{make.L}}

\item{K}{is list of length m where each element represents the movement (diffusion) kernel for that behavior state \code{\link{gausskern}}}

\item{P}{a matrix of dimensions m x m representing the probability for transitions between states.}

\item{m}{is integer indicating the number of desired behavior states. Currently limited to max 2.}

\item{maskL}{is logical indicating whether to mask the input L layer. Default is FALSE. See
\code{mask.L} for details.}

\item{bound.thr}{is numeric indicating the percent threshold that is added 
and subtracted from the bounding box of the filter output from the 
previous day before masking. Default is .05 (5 percent).}

\item{minBounds}{is size (in degrees) of the minimum bounding box around the 
previous days filter prediction that L data within that box will be 
included. Outside this box (centered on t-1 filter prediction), L will be 
masked out.}
}
\value{
a list: list(phi = phi, pred = pred, psi = psi) where \itemize{ \item
  phi. is the probability for each state at each time step \item pred. is
  .... \item psi. is.... }
}
\description{
HMM filter functions
}
\examples{
\dontrun{
# Not run as function relies on large arrays of likelihoods
# RUN THE FILTER STEP
f <- hmm.filter(g, L, K = list(K1, K2), P = P.final, m = 2)
nllf <- -sum(log(f$psi[f$psi>0])) # negative log-likelihood

}

}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290. doi: 10.1111/j.1600-0706.2011.19044.x
}
