% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregular_ncToRaster.r
\name{irregular_ncToRaster}
\alias{irregular_ncToRaster}
\title{Coerce irregular grid as nc to raster}
\usage{
irregular_ncToRaster(fname, varid)
}
\arguments{
\item{fname}{is input filename of the .nc file}
}
\value{
a raster version of the input .nc file coerced to regular grid
}
\description{
Main purpose of this function is for converting bathymetry data that is typically acquired as an irregular grid
}
