% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.wc.r
\name{read.wc}
\alias{read.wc}
\title{Read and format tag data}
\usage{
read.wc(ptt, filename, tag, pop, type = "sst", verbose = FALSE)
}
\arguments{
\item{ptt}{is individual ID number}

\item{filename}{is path to the file where your data lives}

\item{tag}{is POSIXct object of the tagging date}

\item{pop}{is POSIXct object of the pop-up date}

\item{type}{is character indicating which type of data to read. Choices are 
'sst', 'pdt', 'light' corresponding to those data files output from WC Data
Portal}

\item{verbose}{is logical indicating whether a verbose output with more details on the loaded files is desired. Default is FALSE.}
}
\value{
a list containing: the data read as a data.frame and a date vector of
  unique dates in that data
}
\description{
\code{read.wc} reads and formats tag data output from Wildlife Computers Data
Portal
}
\examples{
\dontrun{
# example data loading using example data in the package
sstFile <- system.file("extdata", "141259-SST.csv", package = "HMMoce")
tag.sst <- read.wc(ptt, sstFile, type = 'sst', tag=tag, pop=pop)
}
}
