% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.wc.r
\name{read.wc}
\alias{read.wc}
\title{Read and format tag data}
\usage{
read.wc(ptt, wd = getwd(), tag, pop, type = "sst", verbose = TRUE)
}
\arguments{
\item{ptt}{is individual ID number}

\item{wd}{is the directory where your data lives}

\item{tag}{is POSIXct object of the tagging date}

\item{pop}{is POSIXct object of the pop-up date}

\item{type}{is character indicating which type of data to read. Choices are 
'sst', 'pdt', 'light' corresponding to those data files output from WC Data
Portal}
}
\value{
a list containing: the data read as a data.frame and a date vector of
  unique dates in that data
}
\description{
\code{read.wc} reads and formats tag data output from Wildlife Computers Data
Portal
}
