% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.r
\name{repmat}
\alias{repmat}
\title{Repeat your matrix?}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{is matrix you want to repeat}

\item{m}{is output row dimension}

\item{n}{is output col dimension}
}
\value{
repeated matrix
}
\description{
Repeat your matrix?
}
\references{
Pedersen MW, Patterson TA, Thygesen UH, Madsen H (2011)
  Estimating animal behavior and residency from movement data. Oikos
  120:1281-1290.
}
