% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.grid.par.r
\name{resample.grid.par}
\alias{resample.grid.par}
\title{Resample likelihood rasters to common resolution/extent}
\usage{
resample.grid.par(L.rasters, L.resol, ncores = parallel::detectCores())
}
\arguments{
\item{L.rasters}{list of individual likelihood rasters generated by calc
functions}

\item{L.resol}{raster or raster brick indicating desired output resolution
of all likelihood rasters.}

\item{ncores}{is integer indicating number of cores used in this parallel computation. Defaults to using a detection function that chooses cores for you.}
}
\value{
a list of all resampled likelihood rasters and g, the common grid
}
\description{
Resample likelihood rasters to common resolution/extent
}
