% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.grid.R
\name{resample.grid}
\alias{resample.grid}
\title{Resample likelihood rasters to common resolution/extent}
\usage{
resample.grid(L.rasters, L.res, mle.res = 0.75, bound = NULL)
}
\arguments{
\item{L.rasters}{list of individual likelihood rasters generated by calc 
functions}

\item{L.res}{raster or raster brick indicating desired output resolution of 
all likelihood rasters.}

\item{mle.res}{is desired resolution (in degrees) of the coarse output grid
to be used in parameter estimation.}

\item{bound}{is a raster extent object that can be passed in order to 
restrict the input raster likelihoods to a certain spatial domain. Default 
is null and thus uses the full extent of input rasters. For more 
information on the extent object see \code{raster::extent}}
}
\value{
a list of all resampled likelihood rasters and g, the common grid
}
\description{
Resample likelihood rasters to common resolution/extent
}
