% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyLocs.r
\name{simplifyLocs}
\alias{simplifyLocs}
\title{Simplifies locations to one per day}
\usage{
simplifyLocs(locations, loc.dts)
}
\arguments{
\item{locations}{is data frame from -Locations file output from DAP/Tag Portal for WC tags and 
contains GPS, Argos, and GPE locations as applicable.}

\item{loc.dts}{is vector of dates from locations input}
}
\value{
simplified (one per day) version of locations input
}
\description{
Simplifies locations to one per day
}
